[Decomp][Unit] Cohesion split: extract stat-modifier system to UnitStatModifier.cpp
