+++ src/game/Object/Unit.cpp
@@ -6341,277 +6341,6 @@ uint32 Unit::GetCreatureType() const
     }
 }
 
-/*#######################################
-########                         ########
-########       STAT SYSTEM       ########
-########                         ########
-#######################################*/
-
-bool Unit::HandleStatModifier(UnitMods unitMod, UnitModifierType modifierType, float amount, bool apply)
-{
-    if (unitMod >= UNIT_MOD_END || modifierType >= MODIFIER_TYPE_END)
-    {
-        sLog.outError("ERROR in HandleStatModifier(): nonexistent UnitMods or wrong UnitModifierType!");
-        return false;
-    }
-
-    float val = 1.0f;
-
-    switch (modifierType)
-    {
-        case BASE_VALUE:
-        case TOTAL_VALUE:
-            m_auraModifiersGroup[unitMod][modifierType] += apply ? amount : -amount;
-            break;
-        case BASE_PCT:
-        case TOTAL_PCT:
-            if (amount <= -100.0f)                          // small hack-fix for -100% modifiers
-            {
-                amount = -200.0f;
-            }
-
-            val = (100.0f + amount) / 100.0f;
-            m_auraModifiersGroup[unitMod][modifierType] *= apply ? val : (1.0f / val);
-            break;
-
-        default:
-            break;
-    }
-
-    if (!CanModifyStats())
-    {
-        return false;
-    }
-
-    switch (unitMod)
-    {
-        case UNIT_MOD_STAT_STRENGTH:
-        case UNIT_MOD_STAT_AGILITY:
-        case UNIT_MOD_STAT_STAMINA:
-        case UNIT_MOD_STAT_INTELLECT:
-        case UNIT_MOD_STAT_SPIRIT:         UpdateStats(GetStatByAuraGroup(unitMod));  break;
-
-        case UNIT_MOD_ARMOR:               UpdateArmor();           break;
-        case UNIT_MOD_HEALTH:              UpdateMaxHealth();       break;
-
-        case UNIT_MOD_MANA:
-        case UNIT_MOD_RAGE:
-        case UNIT_MOD_FOCUS:
-        case UNIT_MOD_ENERGY:
-        case UNIT_MOD_RUNE:
-        case UNIT_MOD_RUNIC_POWER:         UpdateMaxPower(GetPowerTypeByAuraGroup(unitMod)); break;
-
-        case UNIT_MOD_RESISTANCE_HOLY:
-        case UNIT_MOD_RESISTANCE_FIRE:
-        case UNIT_MOD_RESISTANCE_NATURE:
-        case UNIT_MOD_RESISTANCE_FROST:
-        case UNIT_MOD_RESISTANCE_SHADOW:
-        case UNIT_MOD_RESISTANCE_ARCANE:   UpdateResistances(GetSpellSchoolByAuraGroup(unitMod)); break;
-
-        case UNIT_MOD_ATTACK_POWER:        UpdateAttackPowerAndDamage();         break;
-        case UNIT_MOD_ATTACK_POWER_RANGED: UpdateAttackPowerAndDamage(true);     break;
-
-        case UNIT_MOD_DAMAGE_MAINHAND:     UpdateDamagePhysical(BASE_ATTACK);    break;
-        case UNIT_MOD_DAMAGE_OFFHAND:      UpdateDamagePhysical(OFF_ATTACK);     break;
-        case UNIT_MOD_DAMAGE_RANGED:       UpdateDamagePhysical(RANGED_ATTACK);  break;
-
-        default:
-            break;
-    }
-
-    return true;
-}
-
-/**
- * @brief Gets a raw stored modifier value for a unit stat bucket.
- *
- * @param unitMod The unit modifier group.
- * @param modifierType The modifier type within the group.
- * @return The stored modifier value.
- */
-float Unit::GetModifierValue(UnitMods unitMod, UnitModifierType modifierType) const
-{
-    if (unitMod >= UNIT_MOD_END || modifierType >= MODIFIER_TYPE_END)
-    {
-        sLog.outError("attempt to access nonexistent modifier value from UnitMods!");
-        return 0.0f;
-    }
-
-    if (modifierType == TOTAL_PCT && m_auraModifiersGroup[unitMod][modifierType] <= 0.0f)
-    {
-        return 0.0f;
-    }
-
-    return m_auraModifiersGroup[unitMod][modifierType];
-}
-
-/**
- * @brief Computes the fully modified value of a primary stat.
- *
- * @param stat The stat to evaluate.
- * @return The resulting total stat value.
- */
-float Unit::GetTotalStatValue(Stats stat) const
-{
-    UnitMods unitMod = UnitMods(UNIT_MOD_STAT_START + stat);
-
-    if (m_auraModifiersGroup[unitMod][TOTAL_PCT] <= 0.0f)
-    {
-        return 0.0f;
-    }
-
-    // value = ((base_value * base_pct) + total_value) * total_pct
-    float value  = m_auraModifiersGroup[unitMod][BASE_VALUE] + GetCreateStat(stat);
-    value *= m_auraModifiersGroup[unitMod][BASE_PCT];
-    value += m_auraModifiersGroup[unitMod][TOTAL_VALUE];
-    value *= m_auraModifiersGroup[unitMod][TOTAL_PCT];
-
-    return value;
-}
-
-/**
- * @brief Computes the fully modified aura contribution for a unit modifier group.
- *
- * @param unitMod The unit modifier group.
- * @return The resulting total modified value.
- */
-float Unit::GetTotalAuraModValue(UnitMods unitMod) const
-{
-    if (unitMod >= UNIT_MOD_END)
-    {
-        sLog.outError("attempt to access nonexistent UnitMods in GetTotalAuraModValue()!");
-        return 0.0f;
-    }
-
-    if (m_auraModifiersGroup[unitMod][TOTAL_PCT] <= 0.0f)
-    {
-        return 0.0f;
-    }
-
-    float value  = m_auraModifiersGroup[unitMod][BASE_VALUE];
-    value *= m_auraModifiersGroup[unitMod][BASE_PCT];
-    value += m_auraModifiersGroup[unitMod][TOTAL_VALUE];
-    value *= m_auraModifiersGroup[unitMod][TOTAL_PCT];
-
-    return value;
-}
-
-/**
- * @brief Maps a resistance unit modifier group to its spell school.
- *
- * @param unitMod The unit modifier group.
- * @return The associated spell school.
- */
-SpellSchools Unit::GetSpellSchoolByAuraGroup(UnitMods unitMod) const
-{
-    SpellSchools school = SPELL_SCHOOL_NORMAL;
-
-    switch (unitMod)
-    {
-        case UNIT_MOD_RESISTANCE_HOLY:     school = SPELL_SCHOOL_HOLY;          break;
-        case UNIT_MOD_RESISTANCE_FIRE:     school = SPELL_SCHOOL_FIRE;          break;
-        case UNIT_MOD_RESISTANCE_NATURE:   school = SPELL_SCHOOL_NATURE;        break;
-        case UNIT_MOD_RESISTANCE_FROST:    school = SPELL_SCHOOL_FROST;         break;
-        case UNIT_MOD_RESISTANCE_SHADOW:   school = SPELL_SCHOOL_SHADOW;        break;
-        case UNIT_MOD_RESISTANCE_ARCANE:   school = SPELL_SCHOOL_ARCANE;        break;
-
-        default:
-            break;
-    }
-
-    return school;
-}
-
-/**
- * @brief Maps a stat unit modifier group to its primary stat.
- *
- * @param unitMod The unit modifier group.
- * @return The associated stat.
- */
-Stats Unit::GetStatByAuraGroup(UnitMods unitMod) const
-{
-    Stats stat = STAT_STRENGTH;
-
-    switch (unitMod)
-    {
-        case UNIT_MOD_STAT_STRENGTH:    stat = STAT_STRENGTH;      break;
-        case UNIT_MOD_STAT_AGILITY:     stat = STAT_AGILITY;       break;
-        case UNIT_MOD_STAT_STAMINA:     stat = STAT_STAMINA;       break;
-        case UNIT_MOD_STAT_INTELLECT:   stat = STAT_INTELLECT;     break;
-        case UNIT_MOD_STAT_SPIRIT:      stat = STAT_SPIRIT;        break;
-
-        default:
-            break;
-    }
-
-    return stat;
-}
-
-/**
- * @brief Maps a power unit modifier group to its power type.
- *
- * @param unitMod The unit modifier group.
- * @return The associated power type.
- */
-Powers Unit::GetPowerTypeByAuraGroup(UnitMods unitMod) const
-{
-    switch (unitMod)
-    {
-        case UNIT_MOD_MANA:       return POWER_MANA;
-        case UNIT_MOD_RAGE:       return POWER_RAGE;
-        case UNIT_MOD_FOCUS:      return POWER_FOCUS;
-        case UNIT_MOD_ENERGY:     return POWER_ENERGY;
-        case UNIT_MOD_RUNE:       return POWER_RUNE;
-        case UNIT_MOD_RUNIC_POWER: return POWER_RUNIC_POWER;
-        default:                  return POWER_MANA;
-    }
-}
-
-/**
- * @brief Computes the effective attack power for an attack type.
- *
- * @param attType The attack type to evaluate.
- * @return The resulting attack power value.
- */
-float Unit::GetTotalAttackPowerValue(WeaponAttackType attType) const
-{
-    if (attType == RANGED_ATTACK)
-    {
-        int32 ap = GetInt32Value(UNIT_FIELD_RANGED_ATTACK_POWER) + GetInt32Value(UNIT_FIELD_RANGED_ATTACK_POWER_MOD_POS);
-        if (ap < 0)
-        {
-            return 0.0f;
-        }
-        return ap * (1.0f + GetFloatValue(UNIT_FIELD_RANGED_ATTACK_POWER_MULTIPLIER));
-    }
-    else
-    {
-        int32 ap = GetInt32Value(UNIT_FIELD_ATTACK_POWER) + GetInt32Value(UNIT_FIELD_ATTACK_POWER_MOD_POS);
-        if (ap < 0)
-        {
-            return 0.0f;
-        }
-        return ap * (1.0f + GetFloatValue(UNIT_FIELD_ATTACK_POWER_MULTIPLIER));
-    }
-}
-
-/**
- * @brief Gets the stored weapon damage range value for an attack type.
- *
- * @param attType The attack type.
- * @param type The minimum or maximum range selector.
- * @return The stored damage range value.
- */
-float Unit::GetWeaponDamageRange(WeaponAttackType attType , WeaponDamageRange type) const
-{
-    if (attType == OFF_ATTACK && !haveOffhandWeapon())
-    {
-        return 0.0f;
-    }
-
-    return m_weaponDamage[attType][type];
-}
-
 /**
  * @brief Sets the unit level and updates group state when needed.
  *
